tests/wpt/web-platform-tests/media-source/dedicated-worker/mediasource-worker-objecturl.html

<!DOCTYPE html>
<html>
<title>Test MediaSource object and objectUrl creation and revocation, with MediaSource in dedicated worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="mediasource-message-util.js"></script>
<script>

async_test(t => {
  // Fail fast if MSE-in-Workers is not supported.
  assert_true(MediaSource.hasOwnProperty("canConstructInDedicatedWorker"), "MediaSource hasOwnProperty 'canConstructInDedicatedWorker'");
  assert_true(MediaSource.canConstructInDedicatedWorker, "MediaSource.canConstructInDedicatedWorker");

  let worker = new Worker("mediasource-worker-play.js");
  worker.onmessage = t.step_func(e => {
    let subject = e.data.subject;
    assert_true(subject != undefined, "message must have a subject field");
    switch (subject) {
      case messageSubject.ERROR:
        assert_unreached("Worker error: " + e.data.info);
        break;
      case messageSubject.OBJECT_URL:
        const url = e.data.info;
        assert_true(url.match(/^blob:.+/) != null);
        URL.revokeObjectURL(url);
        // TODO(crbug.com/1196040): Revoking MediaSource objectURLs on thread
        // that didn't create them is at best a no-op. This test case is
        // possibly obsolete.
        t.done();
        break;
      default:
        assert_unreached("Unexpected message subject: " + subject);

    }
  });
}, "Test main context revocation of DedicatedWorker MediaSource object URL");

if (MediaSource.hasOwnProperty("canConstructInDedicatedWorker") && MediaSource.canConstructInDedicatedWorker === true) {
  // If implementation claims support for MSE-in-Workers, then fetch and run
  // some tests directly in another dedicated worker and get their results
  // merged into those from this page.
  fetch_tests_from_worker(new Worker("mediasource-worker-objecturl.js"));
} else {
  // Otherwise, fetch and run a test that verifies lack of support of
  // MediaSource construction in another dedicated worker.
  fetch_tests_from_worker(new Worker("mediasource-worker-must-fail-if-unsupported.js"));
}

</script>
</html>
